[term] Move some code out of term.c in order to reduce its size
